Add before_disconnect callback to Connection
Added a simple callback to Connection. This callback will be called from the `disconnect/2` function on
MyXQL.Connection, enabling users to do extra stuff just before the connection is disconnected (for instance
send a COM_QUIT as per the MySQL connection conventions).

The callback gets the state of the Connection as its only argument. Part
of this state is `state.client`, which can be used to send the
aforementioned COM_QUIT:

in `config/dev.exs`
```elixir
config :my_app, MyApp.Repo,
  username: "...",
  port: 3306,
  before_disconnect_callback: fn state ->
    MyXQL.Client.com_quit(state.client)
  end
```

This approach where we use a callback allows for opt-in from developers
and does not change any behavior unless a callback is actually implemented.